Decouple presentation from data and logic
- Components now come in the form of containers, enhancers (HoCs) and presenters (view logic - functional)
