+++ packages/docs-site/src/components/presenters/content-box/index.js
